+++ TP-POO1920-RaceCars/DGV.cpp
@@ -4,7 +4,7 @@
 
 DGV::DGV(const DGV& src)
 {
-	int vectorSize = src.carList.size();
+	int vectorSize = static_cast<int>(src.carList.size());
 	if (vectorSize > 0)
 	{
 		for (int i = 0; i < vectorSize; i++)
@@ -13,7 +13,7 @@ DGV::DGV(const DGV& src)
 		}
 	}
 
-	vectorSize = src.pilotList.size();
+	vectorSize = static_cast<int>(src.pilotList.size());
 
 	if (vectorSize > 0)
 	{
@@ -40,7 +40,7 @@ DGV::DGV(const DGV& src)
 
 DGV::~DGV()
 {
-	int pilotVectorSize = pilotList.size();
+	int pilotVectorSize = static_cast<int>(pilotList.size());
 
 	if (pilotVectorSize > 0)
 		for (int i = pilotVectorSize - 1; i >= 0; i--)
@@ -48,7 +48,7 @@ DGV::~DGV()
 
 	pilotList.clear();
 
-	int carVectorSize = carList.size();
+	int carVectorSize = static_cast<int>(carList.size());
 
 	if (carVectorSize > 0)
 		for(int i = carVectorSize - 1; i >= 0; i--)
@@ -88,7 +88,7 @@ bool DGV::addSurprisePilot(const SurprisePilot& newPilot)
 
 bool DGV::removeCar(const char wantedID)
 {
-	int vectorSize = carList.size();
+	int vectorSize = static_cast<int>(carList.size());
 
 	if (vectorSize > 0)
 	{
@@ -113,7 +113,7 @@ bool DGV::removeCar(const char wantedID)
 
 bool DGV::removePilot(std::string wantedName)
 {	
-	int vectorSize = pilotList.size();
+	int vectorSize = static_cast<int>(pilotList.size());
 
 	if (vectorSize > 0)
 	{
@@ -138,7 +138,7 @@ bool DGV::removePilot(std::string wantedName)
 
 bool DGV::printCars()
 {
-	int vectorSize = carList.size();
+	int vectorSize = static_cast<int>(carList.size());
 	if(vectorSize == 0)
 		return false;
 	for (int i = 0; i < vectorSize; i++)
@@ -148,7 +148,7 @@ bool DGV::printCars()
 
 bool DGV::printPilots()
 {
-	int vectorSize = pilotList.size();
+	int vectorSize = static_cast<int>(pilotList.size());
 	if(vectorSize == 0)
 		return false;
 	for (int i = 0; i < vectorSize; i++)
@@ -158,7 +158,7 @@ bool DGV::printPilots()
 
 Car* DGV::getCarById(char wantedID)
 {
-	int vectorSize = carList.size();
+	int vectorSize = static_cast<int>(carList.size());
 	for(int i = 0; i < vectorSize; i++)
 	{
 		char tmpChar=carList[i]->getID();
@@ -170,7 +170,7 @@ Car* DGV::getCarById(char wantedID)
 
 Pilot* DGV::getPilotByName(std::string wantedName)
 {
-	int vectorSize = pilotList.size();
+	int vectorSize = static_cast<int>(pilotList.size());
 	for(int i = 0; i < vectorSize; i++)
 	{
 		std::string tmpString = pilotList[i]->getName();
@@ -225,7 +225,7 @@ bool DGV::detach(char id)
 
 void DGV::detachAll()
 {
-	int vectorSize = carList.size();
+	int vectorSize = static_cast<int>(carList.size());
 	if (vectorSize > 0)
 	{
 		for (int i = 0; i < vectorSize; i++)
@@ -233,7 +233,7 @@ void DGV::detachAll()
 			carList[i]->detach();
 		}
 	}
-	vectorSize = pilotList.size();
+	vectorSize = static_cast<int>(pilotList.size());
 	if(vectorSize > 0)
 	{
 		for(int i = 0; i < vectorSize; i++)
@@ -246,7 +246,7 @@ void DGV::detachAll()
 std::vector<Car*> DGV::getCarsWithPilots()
 {
 	std::vector<Car*> newList;
-	int vectorSize = carList.size();
+	int vectorSize = static_cast<int>(carList.size());
 
 	for(int i = 0; i < vectorSize; i++)
 	{
@@ -261,7 +261,7 @@ int DGV::getNumberOfCarsWithPilots()
 {
 	int numberOfCarsWithPilots = 0;
 
-	int vectorSize = carList.size();
+	int vectorSize = static_cast<int>(carList.size());
 	
 	for(int i = 0; i < vectorSize;i++)
 	{
@@ -275,7 +275,7 @@ int DGV::getNumberOfCarsWithPilots()
 DGV& DGV::operator=(const DGV& src)
 {
 	//Cleaning the destination
-	int vectorSize = this->pilotList.size();
+	int vectorSize = static_cast<int>(this->pilotList.size());
 
 	if (vectorSize > 0)
 		for (int i = vectorSize - 1; i >= 0; i--)
@@ -283,7 +283,7 @@ DGV& DGV::operator=(const DGV& src)
 
 	this->pilotList.clear();
 
-	vectorSize = this->carList.size();
+	vectorSize = static_cast<int>(this->carList.size());
 
 	if (vectorSize > 0)
 		for (int i = vectorSize - 1; i >= 0; i--)
@@ -293,7 +293,7 @@ DGV& DGV::operator=(const DGV& src)
 
 	//Assignment
 	
-	vectorSize = src.carList.size();
+	vectorSize = static_cast<int>(src.carList.size());
 
 	if (vectorSize > 0)
 	{
@@ -303,7 +303,7 @@ DGV& DGV::operator=(const DGV& src)
 		}
 	}
 
-	vectorSize = src.pilotList.size();
+	vectorSize = static_cast<int>(src.pilotList.size());
 
 	if (vectorSize > 0)
 	{
